Version 1.1
Added caching to decrease the loading time. By default it caches to "$HOME/.cheats" but that can be overriden from .vimrc with:
let g:cheats_cache = "/path/to/your/file"

Also the plugin does not depend on any *nix commands any longer (cut, grep, cat) so it may be possible to run it on other platforms as well.